split parsing.h to multiple headerfiles
